fix(skill): close routing gaps + share the path matcher, fix #anchor handling
Addresses all five findings from the /code-review high re-gate on #250.

CORRECTNESS — two routing rows sent a request to a file that did not carry
the content the request needed. Both are the directional-rule failure this
PR's own extraction rule describes, which is a useful datum: writing the
rule down did not prevent violating it two rows below.

  - Task creation named "tags" but routed only to task-creation.md, whose
    Tag Operations section covers nested-path syntax and tag_manage
    mechanics -- not the vocabulary. The @computer/@phone/@office/
    @waiting-for/@someday taxonomy lives in gtd-methodology.md. Pre-split
    the two were always co-resident. A capture would silently invent a
    context tag and fragment the taxonomy, with nothing surfacing the
    divergence.
  - "Process my inbox" routed to workflow-process-inbox.md +
    task-creation.md. Its clarify step delegates to "@waiting-for" but the
    defer window and structured note format are in gtd-methodology.md,
    which only the weekly-review row pulled in. A waiting-for item would be
    tagged and then lose its follow-up date entirely.

Both rows now name gtd-methodology.md and say what it is wanted for.

CORRECTNESS — the new guard classified `file.md#section` as malformed and
failed CLOSED on a valid, resolvable link. A contributor adding a normal
anchor-qualified cross-reference would get a CI error that reads exactly
like a broken link. normalizeRef now strips the fragment, which also fixes
the reachability check that shared the defect. Verified both directions:
a valid anchored link passes; `references/nonexistent.md#x` still fails.

CLEANUP — the two docs guards carried near-identical private copies of
extractRefs/normalizeRef/classifyRef, already diverged on the path-token
regex (`[^\s`)]+` vs `[^\s`)]*`). Extracted to markdown-path-refs.ts,
parameterized by prefix list and allowed extensions. Unifying on the
zero-or-more form is what lets a bare `references/` directory mention be
recognized; the CLAUDE.md guard's eight fixture tests still pass unchanged,
which is what confirms the unification did not loosen it.

npm run build and lint clean. test:unit 3919 passed / 1 failed -- the sole
failure is productivity-period.test.ts, the pre-existing UTC month-boundary
bug filed as OMN-298, unrelated to this PR and failing on origin/main too.
