+++ docs/skills/omnifocus-assistant/SKILL.md
@@ -42,11 +42,11 @@ needed_, not by which topic it concerns.
 | --------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- | --------------------------------------- |
 | Turning any request into a tool call — the natural-language to `omnifocus_read` / `_write` / `_analyze` mapping                                                                                                                                                                                                                               | `references/intent-to-tool-calls.md`    |
 | GTD guidance asked for, or a workflow needs GTD definitions (waiting-for, someday/maybe, defer-vs-due)                                                                                                                                                                                                                                        | `references/gtd-methodology.md`         |
-| "Process my inbox" — read `references/task-creation.md` too before the Execute step (batch-create has a silent `project` limitation)                                                                                                                                                                                                          | `references/workflow-process-inbox.md`  |
+| "Process my inbox" — also read `references/task-creation.md` before the Execute step (batch-create has a silent `project` limitation) and `references/gtd-methodology.md` for the waiting-for / someday-maybe capture conventions (defer window and note format) the clarify step delegates to                                                | `references/workflow-process-inbox.md`  |
 | "Weekly review" — a full review spans four files: also read `references/gtd-methodology.md`, `references/workflow-process-inbox.md` (step 1 empties the inbox) and `references/interpreting-results.md` (step 12 reads productivity stats)                                                                                                    | `references/workflow-weekly-review.md`  |
 | "What's on for today?" / daily planning                                                                                                                                                                                                                                                                                                       | `references/workflow-daily-planning.md` |
 | "What should I work on now?" / choosing among available tasks                                                                                                                                                                                                                                                                                 | `references/workflow-engage.md`         |
-| Creating or rewriting a task or project — naming, time estimates, tags, sequencing, **repetition / recurrence rules**, project **review intervals**                                                                                                                                                                                           | `references/task-creation.md`           |
+| Creating or rewriting a task or project — naming, time estimates, sequencing, **repetition / recurrence rules**, project **review intervals**; for **which tag to use**, also read `references/gtd-methodology.md` — the context-tag vocabulary lives there, and `task-creation.md` covers only nested-tag syntax and `tag_manage` mechanics  | `references/task-creation.md`           |
 | Non-trivial queries — date ranges, tag combinations, projections, counts, pagination; also **exports** (there is no server-side export tool)                                                                                                                                                                                                  | `references/advanced-queries.md`        |
 | `omnifocus_analyze` in either direction — **read this BEFORE calling `parse_meeting_notes`** (you must structure the items yourself, not paste raw prose), and after any analyze call to read back productivity stats, pattern analysis, judgment detectors; also **stalled / stale projects** (30+ days unchanged) and what to do about them | `references/interpreting-results.md`    |
